Move dicard emitter to discard.go
